SyncMethodOrBlock
Synchronized method will lock its instance object and run the code. In the example above, it locks the object p. The second thread will wait until the lock is release on p. Even though dosomething2 method is independent to do something1, all thread will wait until the lock is released on dosomething1.

This is not a cleaner approach if you have multiple synchronized methods and all are independent to each other.

In such cases, you need to create synchronized block with a lock on some object like above locker1, locker2. So other thread will not wait until do something1 is executed.